refactor: 검색 페이지와 결과 페이지 분리
