+++ client/src/pages/result/img.tsx
@@ -3,7 +3,7 @@ import { GetServerSideProps } from "next";
 import { useRouter } from "next/router";
 import Link from "next/link";
 // import SearchedImage from "../../hooks/Search/searchedImage";
-import SearchedImage from "../../components/Main/DropZone";
+import SearchedImage from "../../components/Main/SearchImageBox";
 
 import { useRecoilValue } from "recoil";
 import { searchedImage } from "../../atom/searchedImage";
@@ -17,41 +17,40 @@ import {
   QueryClientProvider,
 } from "react-query";
 import useSearchImg from "../../hooks/Search/useSearchImg";
-import { useSession } from "next-auth/react";
-import DropZone from "../../components/Main/DropZone";
-import SearchBar from "../../components/Common/SearchBar";
 import TextCard from "../../components/Result/TextCard";
 import LoadingScreen from "../../components/Common/LoadingScreen";
 
 const Img = () => {
-  const { data: session, status } = useSession();
   const router = useRouter();
   const { sbi: random_path } = router.query;
   const PushToResult = () => {
-    router.push(`/search/str?data=${data?.data.equal_rate[0].name}`);
+    router.push(`/result/str?data=${data?.data.equal_rate[0].name}`);
   };
   const image = useRecoilValue(searchedImage);
-  const preview = URL.createObjectURL(image);
   const formData = new FormData();
-  console.log(image);
   formData.append("img", image);
+  const preview = URL.createObjectURL(image) || null;
+  
   const { isLoading, data } = useSearchImg(formData, random_path);
   console.log(data);
   const isValidResult = data?.data.equal_rate[0].rate > 0.7;
   console.log(isValidResult);
+  
   if (isValidResult) {
     PushToResult();
   }
 
   if (isLoading) {
-    return <LoadingScreen />;
+    return (
+      <LoadingScreen />
+    )
   }
 
   return (
     <>
       {!isValidResult && (
         <>
-          <img src={preview} />
+          {preview && <img src={preview} />}
           <p>
             최적의 검색결과를 얻지 못했습니다. 혹시 다음 중에 찾으시는 결과가
             있으신가요?
@@ -60,8 +59,6 @@ const Img = () => {
             <TextCard data={item.name} />
           ))}
           <p>다시 검색하기</p>
-          <SearchBar />
-          <DropZone />
         </>
       )}
     </>
